Create Prophet Runner and results
The notebook has methods to run certain types of statistical analysis
to describe the accuracy of the certain predictions when compared with
the actual data. It particularly implements the MAPE and RMSE metrics.

Additionally, it uses Facebook's Prophet model to predict Covid-19 cases
in Brazil and its states from training data. Then it takes this prediction and analyzes its accuracy with the methods mentioned above (along with the MAD metric
that sklearn has already implemented). Finally, it also
generates a graph showing the Prophet model's predictions and another
graph juxtaposing the predictions with actual case data.